feat: updated gepc proxy

+++ proto/grpc_proxy/scripts/gen_protobuf.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 cd ../../..
-yarn sdk Protobuf
-rmdir proto/grpc_proxy/src/main/proto/sdk/4.0
-yarn sdk GrpcProxy
-rmdir proto/grpc_proxy/src/main/java/com/google/looker/server/sdk/4.0
+yarn gen Protobuf
+if [ -d "proto/grpc_proxy/src/main/proto/sdk/4.0" ]; then
+    rmdir proto/grpc_proxy/src/main/proto/sdk/4.0
+fi
+yarn gen GrpcProxy
+if [ -d "proto/grpc_proxy/src/main/java/com/google/looker/server/sdk/4.0" ]; then
+    rmdir proto/grpc_proxy/src/main/java/com/google/looker/server/sdk/4.0
+fi
 rm proto/grpc_proxy/src/main/java/com/google/looker/server/sdk/LookerModels.java
